Converted float to int using typecast

+++ Java Básico/DisplayResolution/src/main/java/com/mycompany/displayresolution/DisplayResolution.java
@@ -18,6 +18,6 @@ public class DisplayResolution {
         Dimension resolution = Toolkit.getDefaultToolkit().getScreenSize();
         double w = resolution.getWidth();
         double h = resolution.getHeight();
-        System.out.println("Screen size: "+ w +"px x " + h + "px");
+        System.out.println("Screen size: "+ (int)w +"px x " + (int)h + "px");
     }
 }
